Record setFrom functions
Changes
- Renamed Record addAll function to setFrom
- Created setFrom overload with key name
- Interval now has Int constructor

Removals
- Deleted unused Record copyFrom function

+++ mayonez-base/src/main/kotlin/mayonez/math/IntMath.kt
@@ -54,7 +54,7 @@ object IntMath {
      */
     @JvmStatic
     fun clamp(value: Int, min: Int, max: Int): Int {
-        return Interval(min.toFloat(), max.toFloat()).clamp(value.toFloat()).toInt()
+        return Interval(min, max).clamp(value.toFloat()).toInt()
     }
